+++ contracts/upgradeable_contracts/BasicForeignBridge.sol
@@ -10,7 +10,7 @@ import "./BasicTokenBridge.sol";
 contract BasicForeignBridge is EternalStorage, Validatable, BasicTokenBridge {
     using SafeMath for uint256;
     /// triggered when relay of deposit from HomeBridge is complete
-    event RelayedMessage(address recipient, uint value, bytes32 transactionHash);
+    event RelayedMessage(address recipient, uint256 value, bytes32 transactionHash);
     function executeSignatures(uint8[] vs, bytes32[] rs, bytes32[] ss, bytes message) external {
         Message.hasEnoughValidSignatures(message, vs, rs, ss, validatorContract());
         address recipient;
@@ -29,13 +29,13 @@ contract BasicForeignBridge is EternalStorage, Validatable, BasicTokenBridge {
         }
     }
 
-    function onExecuteMessage(address, uint256, bytes32) internal returns(bool);
+    function onExecuteMessage(address, uint256, bytes32) internal returns (bool);
 
     function setRelayedMessages(bytes32 _txHash, bool _status) internal {
         boolStorage[keccak256(abi.encodePacked("relayedMessages", _txHash))] = _status;
     }
 
-    function relayedMessages(bytes32 _txHash) public view returns(bool) {
+    function relayedMessages(bytes32 _txHash) public view returns (bool) {
         return boolStorage[keccak256(abi.encodePacked("relayedMessages", _txHash))];
     }
